requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
pydantic
fastapi-mail
python-dotenv
python-multipart
razorpay